Move creatable auto extractor types to /models
